Implement Storybook useArgs for RadioChoiceField data binding
